+++ NHASoftware/Controllers/ForumControllers/ForumPostsController.cs
@@ -2,12 +2,14 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 using Azure.Core;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
+using Microsoft.CodeAnalysis.VisualBasic.Syntax;
 using Microsoft.EntityFrameworkCore;
 using NHASoftware.Data;
 using NHASoftware.Models;
@@ -28,26 +30,25 @@ namespace NHASoftware.Controllers
             _userManager = userManager;
         }
 
-        // GET: ForumPosts
+        /// <summary>
+        /// GET: ForumPosts
+        /// Returns a list of all ForumPosts
+        /// </summary>
+        /// <returns></returns>
         public async Task<IActionResult> Index()
         {
             var applicationDbContext = _context.ForumPost.Include(f => f.ForumTopic);
             return View(await applicationDbContext.ToListAsync());
         }
 
-        // GET: ForumPosts/Details/5
         /// <summary>
-        /// 
+        /// GET: ForumPosts/Details/5
+        /// Returns forumPost Details view for the supplied id. 
         /// </summary>
         /// <param name="id">The forumId you want to view the details page of.</param>
         /// <returns></returns>
         public async Task<IActionResult> Details(int? id)
         {
-            /*******************************************************************************************************
-            *      GET: ForumPosts/Details
-            *      Returns ForumPost Details View. Uses detailVm
-            *******************************************************************************************************/
-
             if (id == null)
             {
                 return NotFound();
@@ -74,15 +75,16 @@ namespace NHASoftware.Controllers
             return View(detailVm);
         }
 
-        // GET: ForumPosts/Create
+        /// <summary>
+        /// GET: ForumPosts/Create
+        /// Populates some server side information for post. 
+        /// Returns the Create forumPost View.
+        /// </summary>
+        /// <param name="id"></param>
+        /// <returns></returns>
         [Authorize]
         public IActionResult Create(int id)
         {
-            /*******************************************************************************************************
-            *      GET: ForumPosts/Create
-            *      Returns forumPost Create View. Populates some server side information.
-            *******************************************************************************************************/
-
             var forumPost = new ForumPost()
             {
                 UserId = _userManager.GetUserId(HttpContext.User),
@@ -95,16 +97,17 @@ namespace NHASoftware.Controllers
             return View(forumPost);
         }
 
+        /// <summary>
+        /// POST: ForumPosts/Create
+        /// Adds forumPost to the database if the model is valid.
+        /// Increments the Post Count & Thread Count & LatestPost.
+        /// </summary>
+        /// <param name="forumPost">ForumPost object to bind values too</param>
+        /// <returns></returns>
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Title,ForumText,CreationDate, UserId, ForumTopicId, CommentCount")] ForumPost forumPost)
         {
-            /*******************************************************************************************************
-             *      POST: ForumPosts/Create
-             *      Adds forumPost to database if the model is valid.
-             *      Increments the Post Count & Thread Count & latestPost.
-             *******************************************************************************************************/
-
             if (ModelState.IsValid)
             {
                 var topic = await _context.ForumTopics.FirstAsync(c => c.Id == forumPost.ForumTopicId);
@@ -120,7 +123,12 @@ namespace NHASoftware.Controllers
             return View(forumPost);
         }
 
-        // GET: ForumPosts/Edit/5
+        /// <summary>
+        /// GET: ForumPosts/Edit/5
+        /// Checks if the id parameter forumPost exist. Returns the forumPost edit view.
+        /// </summary>
+        /// <param name="id">id of the forumPost to be edited. </param>
+        /// <returns></returns>
         public async Task<IActionResult> Edit(int? id)
         {
             /*******************************************************************************************************
@@ -139,21 +147,30 @@ namespace NHASoftware.Controllers
             {
                 return NotFound();
             }
-            ViewData["ForumTopicId"] = new SelectList(_context.ForumTopics, "Id", "Id", forumPost.ForumTopicId);
-            ViewData["reffer"] = Request.Headers["Referer"].ToString();
-            return View(forumPost);
+
+            if (forumPost.UserId == User.FindFirstValue(ClaimTypes.NameIdentifier) || IsUserForumAdmin())
+            {
+                ViewData["ForumTopicId"] = new SelectList(_context.ForumTopics, "Id", "Id", forumPost.ForumTopicId);
+                ViewData["reffer"] = Request.Headers["Referer"].ToString();
+                return View(forumPost);
+            }
+            else
+            {
+                return Unauthorized();
+            }
         }
 
-        // POST: ForumPosts/Edit/5
+        /// <summary>
+        /// POST: ForumPosts/Edit/5
+        /// Updates the forumPost in database.
+        /// </summary>
+        /// <param name="id">forumPost Id</param>
+        /// <param name="forumPost">Post object to bind form values too</param>
+        /// <returns></returns>
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,Title,ForumText,CreationDate,UserId,ForumTopicId, LikeCount")] ForumPost forumPost)
         {
-            /*******************************************************************************************************
-            *      POST: ForumPosts/Edit/3
-            *      Updates the post in database.
-            *******************************************************************************************************/
-
             if (id != forumPost.Id)
             {
                 return NotFound();
@@ -161,37 +178,42 @@ namespace NHASoftware.Controllers
 
             if (ModelState.IsValid)
             {
-                try
+                var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+
+                if (userId == forumPost.UserId || IsUserForumAdmin())
                 {
-                    _context.Update(forumPost);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!ForumPostExists(forumPost.Id))
+                    try
                     {
-                        return NotFound();
+                        _context.Update(forumPost);
+                        await _context.SaveChangesAsync();
                     }
-                    else
+                    catch (DbUpdateConcurrencyException)
                     {
-                        throw;
+                        if (!ForumPostExists(forumPost.Id))
+                        {
+                            return NotFound();
+                        }
+                        else
+                        {
+                            throw;
+                        }
                     }
+                    return RedirectToAction("Details", new {id});
                 }
-                return RedirectToAction("Details", new {id});
             }
             ViewData["ForumTopicId"] = new SelectList(_context.ForumTopics, "Id", "Id", forumPost.ForumTopicId);
             return View(forumPost);
         }
 
-        // GET: ForumPosts/Delete/5
+        /// <summary>
+        /// GET: ForumPosts/Delete/5
+        /// Checks if the id exist. Verifies the user trying to delete post is admin or associated user. Returns DeleteConfirmed View.
+        /// </summary>
+        /// <param name="id"></param>
+        /// <returns></returns>
         [Authorize]
         public async Task<IActionResult> Delete(int? id)
         {
-            /**********************************************************************************************************************************
-            *      GET: ForumPosts/Delete/3
-            *      Checks is id exists. Verifies the user trying to delete post is same as post. Returns DeleteConfirmed View
-            ***********************************************************************************************************************************/
-
             if (id == null)
             {
                 return NotFound();
@@ -205,7 +227,8 @@ namespace NHASoftware.Controllers
             {
                 return NotFound();
             }
-            else if (forumPost.UserId != _userManager.GetUserId(HttpContext.User))
+
+            if (forumPost.UserId != _userManager.GetUserId(HttpContext.User) && !IsUserForumAdmin())
             {
                 return RedirectToAction("Details", "ForumPosts", new {id = forumPost.Id});
             }
@@ -214,28 +237,69 @@ namespace NHASoftware.Controllers
             return View(forumPost);
         }
 
-        // POST: ForumPosts/Delete/5
+        /// <summary>
+        /// POST: ForumPosts/Delete/5
+        /// Deletes the ForumPost form database.
+        /// </summary>
+        /// <param name="id">forumPost id to delete</param>
+        /// <returns></returns>
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            /**********************************************************************************************************************************
-            *      POST: ForumPosts/Delete/3
-            *      Deletes post from database.
-            ***********************************************************************************************************************************/
-
             var forumPost = await _context.ForumPost.FindAsync(id);
+            var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
 
-            var oldPostTopicId = forumPost.ForumTopicId;
-            _context.ForumPost.Remove(forumPost);
-            await _context.SaveChangesAsync();
 
-            return RedirectToAction("Details", "ForumTopics", new{id=oldPostTopicId});
+            if (forumPost != null)
+            {
+                if (userId == forumPost.UserId || IsUserForumAdmin())
+                {
+                    var topic = await _context.ForumTopics.FirstAsync(c => c.Id == forumPost.ForumTopicId);
+
+                    topic.PostCount -= 1;
+                    topic.ThreadCount -= 1;
+                    topic.LastestPost = DateTime.Now;
+
+                    var oldPostTopicId = forumPost.ForumTopicId;
+                    _context.ForumPost.Remove(forumPost);
+                    await _context.SaveChangesAsync();
+                    return RedirectToAction("Details", "ForumTopics", new{id=oldPostTopicId});
+                }
+                else
+                {
+                    return Unauthorized();
+                }
+            }
+            else
+            {
+                return NotFound();
+            }
         }
 
+        
+        /// <summary>
+        /// Checks if the forumPost exist using the id parameter
+        /// </summary>
+        /// <param name="id">ForumPost id to check existence</param>
+        /// <returns>Returns true if ForumPost exist.</returns>
         private bool ForumPostExists(int id)
         {
             return _context.ForumPost.Any(e => e.Id == id);
+        }
+
+        /// <summary>
+        /// Checks if the current logged in user is admin or forum admin
+        /// </summary>
+        /// <returns>Returns Bool if logged in user IS admin or forum admin</returns>
+        private bool IsUserForumAdmin()
+        {
+            if (User.IsInRole("admin") || User.IsInRole("forum admin"))
+            {
+                return true;
+            }
+
+            return false;
         }
     }
 }
